requirements.txt
streamlit==1.14.0
pandas==1.5.0
mysql-connector-python==8.0.28
requests==2.26.0